ref(autofix): Split AutofixCards file into multiple files
This file has gotten quite large with the different types of cards possible.
It's time to split it into multiple smaller files.
