+++ static/app/components/events/autofix/v3/codeChangesCard.tsx
@@ -1,0 +1,128 @@
+import {Fragment, useMemo} from 'react';
+
+import {Button} from '@sentry/scraps/button';
+import {Flex} from '@sentry/scraps/layout';
+import {Text} from '@sentry/scraps/text';
+
+import {
+  collectPatches,
+  getAutofixArtifactFromSection,
+  isCodeChangesArtifact,
+  type AutofixSection,
+  type useExplorerAutofix,
+} from 'sentry/components/events/autofix/useExplorerAutofix';
+import {ArtifactCard} from 'sentry/components/events/autofix/v3/artifactCard';
+import {ArtifactDetails} from 'sentry/components/events/autofix/v3/artifactDetails';
+import {ArtifactLoadingDetails} from 'sentry/components/events/autofix/v3/artifactLoadingDetails';
+import {artifactToMarkdown} from 'sentry/components/events/autofix/v3/utils';
+import {IconCode} from 'sentry/icons/iconCode';
+import {IconRefresh} from 'sentry/icons/iconRefresh';
+import {t, tn} from 'sentry/locale';
+import {useCopyToClipboard} from 'sentry/utils/useCopyToClipboard';
+import {FileDiffViewer} from 'sentry/views/seerExplorer/components/fileDiffViewer';
+
+interface CodeChangesCardProps {
+  autofix: ReturnType<typeof useExplorerAutofix>;
+  section: AutofixSection;
+}
+
+export function CodeChangesCard({autofix, section}: CodeChangesCardProps) {
+  const artifact = useMemo(() => {
+    const sectionArtifact = getAutofixArtifactFromSection(section);
+    return isCodeChangesArtifact(sectionArtifact) ? sectionArtifact : null;
+  }, [section]);
+
+  const {copy} = useCopyToClipboard();
+  const markdown = useMemo(
+    () => (artifact ? artifactToMarkdown(artifact) : null),
+    [artifact]
+  );
+
+  const patchesByRepo = useMemo(() => collectPatches(artifact ?? []), [artifact]);
+
+  const summary = useMemo(() => {
+    const reposChanged = patchesByRepo.size;
+
+    const filesChanged = new Set<string>();
+
+    for (const [repo, patchesForRepo] of patchesByRepo.entries()) {
+      for (const patch of patchesForRepo) {
+        filesChanged.add(`${repo}:${patch.patch.path}`);
+      }
+    }
+
+    if (reposChanged === 1) {
+      return tn(
+        '%s file changed in 1 repo',
+        '%s files changed in 1 repo',
+        filesChanged.size
+      );
+    }
+
+    return t('%s files changed in %s repos', filesChanged.size, reposChanged);
+  }, [patchesByRepo]);
+
+  const {runState, startStep} = autofix;
+  const runId = runState?.run_id;
+
+  return (
+    <ArtifactCard
+      icon={<IconCode />}
+      title={t('Code Changes')}
+      onCopy={
+        markdown
+          ? () => copy(markdown, {successMessage: t('Copied to clipboard.')})
+          : undefined
+      }
+    >
+      {section.status === 'processing' ? (
+        <ArtifactLoadingDetails
+          blocks={section.blocks}
+          loadingMessage={t('Implementing changes\u2026')}
+        />
+      ) : (
+        <Fragment>
+          {patchesByRepo.size ? (
+            <Fragment>
+              <Text>{summary}</Text>
+              {[...patchesByRepo.entries()].map(([repo, patches]) => (
+                <ArtifactDetails key={repo}>
+                  <Flex gap="lg">
+                    <Text bold>{t('Repository:')}</Text>
+                    <Text>{repo}</Text>
+                  </Flex>
+                  {patches.map((patch, index) => (
+                    <FileDiffViewer
+                      key={index}
+                      patch={patch.patch}
+                      showBorder
+                      collapsible
+                      defaultExpanded={artifact !== null && artifact.length <= 1}
+                    />
+                  ))}
+                </ArtifactDetails>
+              ))}
+            </Fragment>
+          ) : (
+            <ArtifactDetails>
+              <Text>
+                {t(
+                  'Seer failed to generate a code change. This one is on us. Try running it again.'
+                )}
+              </Text>
+              <div>
+                <Button
+                  priority="primary"
+                  icon={<IconRefresh />}
+                  onClick={() => startStep('code_changes', runId)}
+                >
+                  {t('Re-run')}
+                </Button>
+              </div>
+            </ArtifactDetails>
+          )}
+        </Fragment>
+      )}
+    </ArtifactCard>
+  );
+}
