+++ static/app/components/events/autofix/v3/solutionCard.tsx
@@ -1,0 +1,98 @@
+import {Fragment, useMemo} from 'react';
+
+import {Button} from '@sentry/scraps/button';
+import {Container, Flex} from '@sentry/scraps/layout';
+import {Text} from '@sentry/scraps/text';
+
+import {
+  getAutofixArtifactFromSection,
+  isSolutionArtifact,
+  type AutofixSection,
+  type useExplorerAutofix,
+} from 'sentry/components/events/autofix/useExplorerAutofix';
+import {ArtifactCard} from 'sentry/components/events/autofix/v3/artifactCard';
+import {ArtifactDetails} from 'sentry/components/events/autofix/v3/artifactDetails';
+import {ArtifactLoadingDetails} from 'sentry/components/events/autofix/v3/artifactLoadingDetails';
+import {StyledMarkedText} from 'sentry/components/events/autofix/v3/styled';
+import {artifactToMarkdown} from 'sentry/components/events/autofix/v3/utils';
+import {IconList} from 'sentry/icons/iconList';
+import {IconRefresh} from 'sentry/icons/iconRefresh';
+import {t} from 'sentry/locale';
+import {useCopyToClipboard} from 'sentry/utils/useCopyToClipboard';
+
+interface SolutionCardProps {
+  autofix: ReturnType<typeof useExplorerAutofix>;
+  section: AutofixSection;
+}
+
+export function SolutionCard({autofix, section}: SolutionCardProps) {
+  const artifact = useMemo(() => {
+    const sectionArtifact = getAutofixArtifactFromSection(section);
+    return isSolutionArtifact(sectionArtifact) ? sectionArtifact : null;
+  }, [section]);
+
+  const {copy} = useCopyToClipboard();
+  const markdown = useMemo(
+    () => (artifact ? artifactToMarkdown(artifact) : null),
+    [artifact]
+  );
+  const {runState, startStep} = autofix;
+  const runId = runState?.run_id;
+
+  return (
+    <ArtifactCard
+      icon={<IconList />}
+      title={t('Plan')}
+      onCopy={
+        markdown
+          ? () => copy(markdown, {successMessage: t('Copied to clipboard.')})
+          : undefined
+      }
+    >
+      {section.status === 'processing' ? (
+        <ArtifactLoadingDetails
+          blocks={section.blocks}
+          loadingMessage={t('Formulating a plan\u2026')}
+        />
+      ) : artifact?.data ? (
+        <Fragment>
+          <StyledMarkedText text={artifact.data.one_line_summary} />
+          {artifact.data.steps ? (
+            <ArtifactDetails>
+              <Text bold>{t('Steps to Resolve')}</Text>
+              <Container as="ol" margin="0">
+                {artifact.data.steps.map((step, index) => (
+                  <li key={index}>
+                    <Flex direction="column">
+                      <StyledMarkedText text={step.title} />
+                      <Text size="sm" variant="muted">
+                        {step.description}
+                      </Text>
+                    </Flex>
+                  </li>
+                ))}
+              </Container>
+            </ArtifactDetails>
+          ) : null}
+        </Fragment>
+      ) : (
+        <ArtifactDetails>
+          <Text>
+            {t(
+              'Seer failed to generate a plan. This one is on us. Try running it again.'
+            )}
+          </Text>
+          <div>
+            <Button
+              priority="primary"
+              icon={<IconRefresh />}
+              onClick={() => startStep('solution', runId)}
+            >
+              {t('Re-run')}
+            </Button>
+          </div>
+        </ArtifactDetails>
+      )}
+    </ArtifactCard>
+  );
+}
