+++ static/app/components/events/autofix/v3/styled.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+import {MarkedText} from 'sentry/utils/marked/markedText';
+
+export const StyledMarkedText = styled(MarkedText)`
+  p {
+    margin: 0;
+  }
+`;
